+++ CHANGELOG.md
@@ -6,8 +6,25 @@
 
 ### Changed
 
+## 0.10.5
+
+### Added
+
+### Changed
+
+- Fix Galaxy note in the README.md.
+- Fix interleaved RGB issue in Avivator when setting up viewer initially.
+
+## 0.10.4
+
+### Added
+
+### Changed
+
 - Make OME-TIFF loader use a `guessTileSize` function just like the zarr loaders in order to handle pyramids that are not tiled, just downsampled.
 - Don't use `changeFlags` for `updateState` prop diffing.
+- Fix broken decoder Pool. Remove pool proxy.
+- Replace Webpack & Rollup with Vite for development & build.
 
 ## 0.10.3